move Set to own util/collections package
