+++ follow_along/glide.yaml
@@ -3,6 +3,7 @@ import:
 - package: github.com/Sotera/go_watchman
   subpackages:
   - redis_dispatcher
+  - util/collections
 - package: golang.org/x/net
   subpackages:
   - html
