--- util/time/glide.yaml
@@ -1,2 +1,0 @@
-package: github.com/Sotera/go_watchman/util/time
-import: []
